Hw5 auth

+++ src/services/contacts.js
@@ -1,66 +1,55 @@
 import { ContactsCollection } from "../db/models/contact.js";
 import { calculatePaginationData } from "../utils/calculatePaginationData.js";
-import { SORT_ORDER } from "../constans/index.js";
-import { parseIsFavourite } from "../utils/parseFilterParams.js";
+import { SORT_ORDER } from '../constants/index.js';
 
-export const getAllContacts = async ( page=1, perPage=10, sortBy = 'name', sortOrder = SORT_ORDER.ASC, filter = {}, userId) => {
-   const limit = perPage;
-    const skip = (page - 1) * perPage;
+export const getAllContacts = async ({ page=1, perPage=10, sortOrder = SORT_ORDER.ASC, sortBy = '_id', filter = {}, userId}) => {
 
-    let contactsQuery = ContactsCollection.find({userId}).skip(skip).limit(limit).sort({ [sortBy]: sortOrder });
+  const limit = perPage;
+  const skip = (page - 1) * perPage;
+  const contactsQuery = ContactsCollection.find();
+  
+  if (filter.isFavourite) {
+    contactsQuery.where('isFavourite').equals(filter.isFavourite);
+  };
+   if (filter.contactType) {
+    contactsQuery.where('contactType').equals(filter.contactType);
+  };
 
-      if (filter.isFavourite !== undefined) {
-        const parsedIsFavourite = parseIsFavourite(filter.isFavourite);
-        if (parsedIsFavourite !== undefined) {
-            contactsQuery = contactsQuery.where('isFavourite').equals(parsedIsFavourite);
-        }
-    }
+  contactsQuery.where('userId').equals(userId);
+  
+const [contactsCount, contacts] = await Promise.all([
+    ContactsCollection.find().merge(contactsQuery).countDocuments(),
+    contactsQuery
+      .skip(skip)
+      .limit(limit)
+      .sort({ [sortBy]: sortOrder })
+      .exec(),
+  ]);
 
-  const contactsCount = await ContactsCollection.countDocuments({userId});
-  const contacts = await contactsQuery.exec();
-  const paginationData = calculatePaginationData(contactsCount, limit, page);
 
+  const paginationData = calculatePaginationData(contactsCount, perPage, page);
   return {
     data: contacts,
     ...paginationData,
   };
 };
 
-export const getContactsById = async (contactId, userId) => {
-  const contact = await ContactsCollection.findOne({ _id: contactId, userId });
+export const getContactById = async (userId,contactId) => {
+  const contact = await ContactsCollection.findOne({ _id: contactId, userId });;
   return contact;
 };
 
-export const createContacts = async (payload) => {
+export const createContact = async (payload) => {
   const contact = await ContactsCollection.create(payload);
   return contact;
 };
 
-
-export const deleteContact = async (contactId, userId) => {
-    const contact = await ContactsCollection.findOneAndDelete({
-      _id: contactId,
-      userId
-    });
-
-    return contact;
+export const patchContactById = async (userId, contactId, payload) => {
+  const contact = await ContactsCollection.findOneAndUpdate({ _id: contactId, userId }, payload, {new:true});
+  return contact;
 };
 
-export const updateContact = async ( contactId,userId, payload, options = {}) => {
-  const rawResult = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId, userId},
-    payload,
-    {
-      new: true,
-      includeResultMetadata: true,
-      ...options,
-    }
-  );
-
-  if (!rawResult || !rawResult.value) return null;
-
-  return {
-    contact: rawResult.value,
-    isNew: Boolean(rawResult?.lastErrorObject?.upserted),
-  };
+export const deleteContactById = async (userId,contactId) => {
+  const contact = await ContactsCollection.findOneAndDelete({_id: contactId, userId });
+  return contact;
 };
